Separated server-side page with client-side components by making awrapper
